Inject 'Test::Synopsis' as a develop dependency
Inject 'Test::Synopsis as a 'develop' dependency into the dist.
Thus, users of the plugin will have 'Test::Synopsis' listed in
"dzil listdeps --develop".

Note that this patch breaks Test::Pod::Coverage. The dzil plugin should
be disabled for this distribution.

Next step: make Test::Synopsis mandatory in the test code.

This can avoid strange bugs where the test is skipped because of a user
installation issue. Instead, if the plugin is referenced in dist.ini we
want to enforce that the test runs and succeed.
Here is an example of the kind of bugs that can happen:
   https://github.com/rjbs/Dist-Zilla/issues/283