Generated test code now uses Test::Synopsis without guard
This is a release test. It should not be silently skipped for obscure
reasons (such as Test::Synopsis not being installed).
So the test will now fail if Test::Synopsis is missing. The developer
that may be victim of this has no excuse because Test::Synopsis is in
"dzil listdeps --author" (see previous commit).